+++ xpeapp_android/app/src/main/java/com/xpeho/xpeapp/XpeApp.kt
@@ -1,0 +1,16 @@
+package com.xpeho.xpeapp
+
+import android.app.Application
+import com.xpeho.xpeapp.di.AppModule
+import com.xpeho.xpeapp.di.MainAppModule
+
+class XpeApp: Application() {
+    companion object {
+        lateinit var appModule: AppModule
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        appModule = MainAppModule(appContext = this)
+    }
+}